THIS IS TEMPORARY AND SHOULD STAY ON TOP OF THE BRANCH.
[TMP]Podcasts: change libvlcjni / medialibrary branches

I didn't write any docs anywhere two years ago on how to build the
native part of the podcasts so this will use the proper repos / branches
and patches for libvlcjni, vlc, medialibrary and libvlcpp

Libvlcjni uses my podcasts-4 branch that patches for vlc.
Medialibrary uses my podcasts branch which has one commit on top.

for libvlcpp the medialibrary buildscript from vlc-android is modified
to cherry-pick a commit from my fork, but we could have added a patch to
the medialibrary patch set in buildsystem.