StartActivity: fix shortcuts for video playlists
When starting a playlist from a launcher shortcut, if the app was killed
or left with the back button play, and the user tries to start a
playlist with the first media being a video, playback will start in the
background without starting the video player.

This forces the video player when opening a video file or a playlist
starting with a video.

This opens the MainPage before opening the video player so that when
leaving the video player, if the app started from a Destroyed state, it
will still go through the MainPage instead of just leaving the app.