buildsystem: build mla native library via gradle
The list of architectures can be set via the command line (selectABI
separated by a comma), otherwise it picks the architectures that have a
matching libmedialibrary.a.

If the no architecture is set and vlcLibVariant is set (single architecture build)
it uses that single architecture.

We need to use a recent NDK for arm64/x86_64 to be able to ship a libc++_shared.lib
that is 16k-aligned. But gradle complains if we target API 17 (VLC 3) with
such a new NDK so we silence the warning when building VLC 3.